requirements.txt
pytorch==1.9.0
torchvision==0.10.0
pillow==8.2.0
numpy
matplotlib
tqdm
json
pyyaml
